[BUGFIX] Improve file handling on Windows systems
Calling unlink() on a symlink that points to a directory
fails on Windows. Solution is to use rmdir() instead to
remove the symlink.
Furthermore it fixes a wrong usage of DIRECTORY_SEPARATOR
within the FileHandlingUtility.

Releases: master, 6.2
Resolves: #64409
Reviewed-on: http://review.typo3.org/36359
Tested-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>

+++ typo3/sysext/extensionmanager/Classes/Utility/FileHandlingUtility.php
@@ -264,8 +264,9 @@ class FileHandlingUtility implements \TYPO3\CMS\Core\SingletonInterface {
 	 * @return void
 	 */
 	public function removeDirectory($extDirPath) {
-		$extensionPathWithoutTrailingSlash = rtrim($extDirPath, DIRECTORY_SEPARATOR);
-		if (is_link($extensionPathWithoutTrailingSlash)) {
+		$extDirPath = GeneralUtility::fixWindowsFilePath($extDirPath);
+		$extensionPathWithoutTrailingSlash = rtrim($extDirPath, '/');
+		if (is_link($extensionPathWithoutTrailingSlash) && TYPO3_OS !== 'WIN') {
 			$result = unlink($extensionPathWithoutTrailingSlash);
 		} else {
 			$result = GeneralUtility::rmdir($extDirPath, TRUE);
